python/pyproject.toml
[tool.poetry]
name = "parser"
version = "0.1.0"
description = ""
authors = ["David Liu <david-khala@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
antlr4-python3-runtime = "^4.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
